update the name of *Command and *Response

+++ ehrest/src/main/java/com/everhomes/rest/news/ListNewsBySceneCommand.java
@@ -18,7 +18,7 @@ public class ListNewsBySceneCommand {
 	@NotNull
 	private String sceneToken;
 	private Long pageAnchor;
-	private Long pageSize;
+	private Integer pageSize;
 
 	public String getSceneToken() {
 		return sceneToken;
@@ -36,11 +36,11 @@ public class ListNewsBySceneCommand {
 		this.pageAnchor = pageAnchor;
 	}
 
-	public Long getPageSize() {
+	public Integer getPageSize() {
 		return pageSize;
 	}
 
-	public void setPageSize(Long pageSize) {
+	public void setPageSize(Integer pageSize) {
 		this.pageSize = pageSize;
 	}
