+++ ehrest/src/main/java/com/everhomes/rest/news/ListNewsCommand.java
@@ -21,7 +21,7 @@ public class ListNewsCommand {
 	@NotNull
 	private Long ownerId;
 	private Long pageAnchor;
-	private Long pageSize;
+	private Integer pageSize;
 
 	public String getOwnerType() {
 		return ownerType;
@@ -47,11 +47,11 @@ public class ListNewsCommand {
 		this.pageAnchor = pageAnchor;
 	}
 
-	public Long getPageSize() {
+	public Integer getPageSize() {
 		return pageSize;
 	}
 
-	public void setPageSize(Long pageSize) {
+	public void setPageSize(Integer pageSize) {
 		this.pageSize = pageSize;
 	}
